Cleanup of 'posts' code that has been moved to users
